not quite so inconceivable that the shepherd count would be huge now
git-svn-id: https://qthreads.googlecode.com/svn/trunk@1366 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ include/qthread/qthread.h
@@ -75,7 +75,7 @@ Q_ENDCXX /* */
 
 Q_STARTCXX /* */
 typedef struct qthread_s qthread_t;
-typedef unsigned short qthread_shepherd_id_t;	/* doubt we'll run more than 65k shepherds */
+typedef unsigned int qthread_shepherd_id_t;
 
 /* for convenient arguments to qthread_fork */
 typedef aligned_t(*qthread_f) (qthread_t * me, void *arg);
